+++ pymince/_constants.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+utf_8 = "utf-8"
+empty = object()